initial pre-commit config
soft launch meaning only people with pre-commit tool installed would see this run.
(full launch would entail, running pre-commit on all files and  github action to enforce pre-commit)

If you want to see full effect on files run: `pre-commit run --all-files` I looked the results of this over and changes seem fine.

There's a lot of formatting changes to the html. (but some of the html looks minified so to be expected)
I'd suggest maintainers look over what the changes would be for all the files and the generate final docs.

fixes #260